Draw Circle Function Complete
100% pixel precision. Did NOT use Bresenham's Law for this, as it
redraws over the same pixels numerous times. This draws only once, but
the for loop cycles through 25% extra non drawn pixels, but still is
probably less of an overhaul than the pixel double-ups in Bresenham's,
not to mention 0 variable initializations outside the 2 in the for
loops.

On to sprites now. Then possibly collision detection!