Collision Detection Alpha
Collision detection mostly works.
It is working for any polygon 3+ sides.
It currently has the following bugs.
(1) Since transition is scaled based on how fast the game is running,
the collision detection is working with the unscaled transition amounts,
this means it is not pixel perfect, and can in some cases cause the
sprite to get stuck. Instead I believe its best to scale the velocity in
the beginning of the game loop, thus any work done with it is
consistent, also allowing us to move the CollisionPredict function
outside of the sprite class, in case we need to use it elsewhere.

Double jumping is working although works stupid.

TODO: We need to go over all the collision interseption points, checking
which is the furthest from the transition point. This way we know which
point is deeper in the sprite we are checking, to move it back
correctly.